fix: 当選したポケモンと図鑑に登録されるポケモンがずれるバグを修正
