+++ src/component/project/Display/index.tsx
@@ -1,6 +1,5 @@
 import { useRef } from "react";
 import { usePokemon } from "src/hook/usePokemon";
-import { usePokemonList } from "src/hook/usePokemonList";
 import { useScreen } from "src/hook/useScreen";
 
 import { Display as DisplayPresenter } from "./Display";
@@ -8,8 +7,7 @@ import type { DisplayProps } from "./Display.type";
 
 const Display: React.FC = () => {
   const { screenType } = useScreen();
-  const { pokemon } = usePokemon();
-  const { pokemonList } = usePokemonList();
+  const { pokemon, pokemonList } = usePokemon();
 
   const pokedexRef = useRef(null);
